Opt out of token scope changes
When multi-domain support is enabled, and the user is perceived
to be an admin in the domain this user is defined in,
Horizon when doing most requests to keystone silently replaces
the user's 'project-scoped' token with 'domain-scoped' one.

This however does not work well with some combinations of policies
(e.g. `admin_required: role:admin and is_admin_project:True`),
so operator should be able to opt out of this token scope swap.

This patch adds new setting `OPENSTACK_KEYSTONE_PREFER_DOMAIN_TOKEN`
(default to `True` to keep current behavior), and token scope change
can be disabled by setting it to `False`.

Closes-Bug: #2115353
Change-Id: I25580380a438f9d73dfe3a3de1097911c50202a7
Signed-off-by: Pavlo Shchelokovskyy <shchelokovskyy@gmail.com>